requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pika
python-dotenv
requests
python-jose[cryptography]
